BREAKING: rm object stacking methods
